+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * issue #256: Fix promotion slots drag and drop sorting.
 * issue #258: Add Expires At Field to Article Screen.
 * Add related teaser fields when the node has the `triniti:curator:mixin:has-related-teasers` mixin.
+* Add display title field to image uploader.
 
 
 ## v0.9.2